flask_server/requirements.txt
flask
gunicorn
Flask-Mailman
python-dotenv
SQLAlchemy  
flask_sqlalchemy
PyJWT 
flask-cors
flask_migrate
pyzbar
qrcode


matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1
tqdm>=4.41.0

sympy
mpmath
packaging
pyparsing
cycler
python-dateutil
kiwisolver
six
pytz

pandas>=1.1.4
seaborn>=0.11.0
